Fix font issues.
The quotations didn't fix the unicode issue. It was using the default
font rather than 'Inconsolata-g for Powerline'.

Likewise, the - needed to be doubly escaped to use to correct font.
However this also removed the symbols in the prompt. Therefore falling
back on 'Deja Vu Sans Mono' if they are not available through xft with
the primary font (though can be displayed through gvim strangely).

Fix unicode issues by installing ttf-ms-win8 using the Windows partition
and editing the MD5s for cabari fonts. However not reflected in this
commit.